# Backent/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EV Charging Platform API",
    "description": "API documentation for the EV Charging Platform project.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api"
    }
  ],
  "paths": {
...
    "/getCurrentUser": {
      "get": {
        "summary": "Get details of current logged-in user",
        "operationId": "getCurrentUser",
        "tags": ["User"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/vehicle/add": {
      "post": {
        "summary": "Add a new vehicle",
        "operationId": "addVehicle",
        "tags": ["Vehicle"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "vehicleNumber": {
                    "type": "string"
                  },
                  "vehicleType": {
                    "type": "string"
                  }
                },
                "required": ["vehicleNumber", "vehicleType"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vehicle added successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/v1/vehicle/all": {
      "get": {
        "summary": "Get all vehicles of the logged-in user",
        "operationId": "getVehicles",
        "tags": ["Vehicle"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of vehicles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "vehicleId": {
                        "type": "string"
                      },
                      "vehicleNumber": {
                        "type": "string"
                      },
                      "vehicleType": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/wallet/create": {
      "post": {
        "summary": "Create a new wallet",
        "operationId": "createWallet",
        "tags": ["Wallet"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Wallet created successfully"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
} 